Fix ejabberdctl_SUITE:num_active_users/1
It seems that the test wouldn't work with XMPP user registration,
due to the fact that user creation (or some other effect from another test)
is recorded as last activity.

By artificially setting the last activity first and later changing it
we have more precise control and can verify if the change is correctly
reflected by `num_active_users` command.